refactor: remove redundant RAP row cast

+++ app/Services/Rap/RapTextParser.php
@@ -302,6 +302,6 @@ class RapTextParser
             return null;
         }
 
-        return (int) end($row);
+        return end($row);
     }
 }
